Ep 6-7 done.
This is going to have to be 2 parts.

 - entity.h: main factory for game objects
 - component.h: abstract holding class for data related to game object
 - component_system.h: abstract execution class to make edits to the game
 - component_system_render_dynamic_draw.h:  draw object to screen offseted by the local camera position
 - component_transform: local position of the object with respect to the camera
 - component_render: holding class for basic render information
 - splay_tree.h: basic class implementation for a splay tree
 - rect.h: moved rect out of renderer.h into its own class
 - game.h/game.cpp: holding class for important global variables
 - other name changes to classes now fitting within the EC engine (material, texture, shader, renderer)

+++ opengl_build_rpg_tutorial/entity.h
@@ -54,7 +54,7 @@ public:
 
 	Entity* get_child(std::string str)
 	{
-		return children_.search(std::hash<std::string>{}(str));
+		return children_.search((int)std::hash<std::string>{}(str));
 	}
 
 	std::vector<Entity*> get_child_list()
